backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
redis
rq
playwright
tldextract
python-dotenv
pytest
httpx

